+++ Source/BugsnagConfiguration.h
@@ -29,7 +29,7 @@
 #import "BSG_KSCrashReportWriter.h"
 #import "BugsnagBreadcrumb.h"
 #import "BugsnagEvent.h"
-#import "BugsnagMetaData.h"
+#import "BugsnagMetadata.h"
 
 @class BugsnagBreadcrumbs;
 @class BugsnagUser;
@@ -120,11 +120,11 @@ typedef NSDictionary *_Nullable (^BugsnagBeforeNotifyHook)(
  *  Additional information about the state of the app or environment at the
  *  time the report was generated
  */
-@property(readwrite, retain, nullable) BugsnagMetaData *metaData;
+@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
 /**
  *  Meta-information about the state of Bugsnag
  */
-@property(readwrite, retain, nullable) BugsnagMetaData *config;
+@property(readwrite, retain, nullable) BugsnagMetadata *config;
 /**
  *  Rolling snapshots of user actions leading up to a crash report
  */