+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/HandledErrorOverrideScenario.swift
@@ -23,7 +23,7 @@ class HandledErrorOverrideScenario: Scenario {
             report.errorMessage = "Foo"
             report.errorClass = "Bar"
             report.depth += 2
-            report.metaData["account"] = [
+            report.metadata["account"] = [
                 "items": [400,200]
             ]
         }