+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/OOMScenario.m
@@ -8,9 +8,9 @@
     self.config.shouldAutoCaptureSessions = NO;
     self.config.releaseStage = @"alpha";
     [self.config addBeforeSendBlock:^bool(NSDictionary * _Nonnull rawEventData, BugsnagEvent * _Nonnull report) {
-        NSMutableDictionary *metadata = [report.metaData mutableCopy];
+        NSMutableDictionary *metadata = [report.metadata mutableCopy];
         metadata[@"extra"] = @{ @"shape": @"line" };
-        report.metaData = metadata;
+        report.metadata = metadata;
         return YES;
     }];
     [super startBugsnag];